refactor: move server and utility handlers
